+++ src/graph.h
@@ -89,7 +89,7 @@ struct EdgePair {
 };
 
 // SG = serialized graph, these types are for writing graph to file
-typedef int32_t SGID;
+typedef int64_t SGID;
 typedef EdgePair<SGID> SGEdge;
 typedef int64_t SGOffset;
